requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
asyncio
aiohttp>=3.8.0
pydantic>=2.0.0
typing-extensions>=4.0.0 